Use a size_t for the LOD level.
We use the LOD level as an index into arrays at various times.  Since
we were forcing it to be 64bit, this meant win32 spit out a ton of warnings
about our index being larger than a valid array index and was being implicitly
cast.  This makes it a size_t so 32bit platforms get fewer warnings.

Unfortunately this means we can no longer work with datasets which have
more than 4 billion LODs.  Sorry.  If you have such a dataset, I suggest
stealing a supercomputer from the (distant) future to process it.

+++ IO/uvfMetadata.cpp
@@ -58,7 +58,7 @@ UVFMetadata::UVFMetadata(RasterDataBlock* pVolumeDataBlock,
 
   // we also assume that x,y,z are in the first 3 components and
   // we have no anisotropy (i.e. ulLODLevelCount.size=1)
-  m_iLODLevel = m_pVolumeDataBlock->ulLODLevelCount[0];
+  m_iLODLevel = static_cast<size_t>(m_pVolumeDataBlock->ulLODLevelCount[0]);
   for (size_t i = 0;i<3;i++) {
     m_aOverlap[i] = m_pVolumeDataBlock->ulBrickOverlap[i];
     m_aMaxBrickSize[i] = m_pVolumeDataBlock->ulBrickSize[i];
